fix(video_qc): satisfy py3.10 mypy and no-extras wheel smoke-test
Two CI-only failures the local (py3.13, extras-installed) run couldn't see:

- mypy on Python 3.10 (mypy 1.19.1 + locked numpy stubs) types np.linspace
  as floating[Any], not float64, so _noise_timestamps' return and the noise
  grid passed to bt709_noise_shape failed the strict float64 annotations.
  Pin dtype=np.float64 on both linspace calls.
- The wheel smoke-test runs tests/ with no plotting extra; test_video_qc.py
  imported matplotlib at module top and broke collection. Guard with
  pytest.importorskip("matplotlib") so the module skips cleanly.

+++ src/aind_video_utils/video_qc.py
@@ -501,7 +501,7 @@ def classify_gamma(
 def _noise_timestamps(input_probe: Any, n_frames: int) -> NDArray[np.float64]:
     """``n_frames`` timestamps spread across the source (fallback to a 60 s guess)."""
     duration = get_duration_seconds(input_probe) or 60.0
-    return np.linspace(0.05 * duration, 0.95 * duration, max(n_frames, 3))
+    return np.linspace(0.05 * duration, 0.95 * duration, max(n_frames, 3), dtype=np.float64)
 
 
 def transcode_qc_figure(
@@ -623,7 +623,7 @@ def transcode_qc_figure(
     ax_noise.plot(mean[~ok], variance[~ok], "o", color="0.78", ms=3)
     ax_noise.plot(mean[ok], variance[ok], "o", color="0.25", ms=3)
     if int(ok.sum()) >= 3:
-        grid = np.linspace(float(mean[ok].min()), float(mean[ok].max()), 60)
+        grid = np.linspace(float(mean[ok].min()), float(mean[ok].max()), 60, dtype=np.float64)
         a_mean, a_var = float(verdict["anchor_mean"]), float(verdict["anchor_var"])
         lin_curve = a_var * (grid / a_mean)
         gam_curve = a_var * bt709_noise_shape(grid) / float(bt709_noise_shape(np.array([a_mean]))[0])
